+++ packages/engine/src/wwa_main.ts
@@ -29,13 +29,13 @@ import { Player } from "./wwa_parts_player";
 import { Monster } from "./wwa_monster";
 import { ObjectMovingDataManager } from "./wwa_motion";
 import {
-    MessageWindow, MosterWindow, WWASaveData,  ScoreWindow, MessageInfo, Macro, parseMacro
+    MessageWindow, MosterWindow,  ScoreWindow, MessageInfo, Macro, parseMacro
 } from "./wwa_message";
 import { BattleEstimateWindow } from "./wwa_estimate_battle";
 import { PasswordWindow, Mode } from "./wwa_password_window";
 import { inject } from "./wwa_inject_html";
 import { ItemMenu } from "./wwa_item_menu";
-import { WWACompress } from "./wwa_psave";
+import { WWACompress, WWASave } from "./wwa_save";
 import { WWAWebAudio, WWAAudioElement, WWAAudio } from "./wwa_audio";
 
 let wwa: WWA;
@@ -97,12 +97,13 @@ export class WWA {
     private _statusPressCounter: Status; // ステータス型があるので、アニメーション残りカウンタもこれで代用しまぁす。
     private _battleEstimateWindow: BattleEstimateWindow;
     private _passwordWindow: PasswordWindow;
+    private _wwaSave: WWASave;
 
     private _stopUpdateByLoadFlag: boolean;
     private _isURLGateEnable: boolean;
     private _loadType: LoadType;
     private _restartData: WWAData;
-    public _checkOriginalMapString: string; 
+    public checkOriginalMapString: string; 
     private _prevFrameEventExected: boolean;
 
     private _reservedMoveMacroTurn: number; // $moveマクロは、パーツマクロの中で最後に効果が現れる。実行されると予約として受け付け、この変数に予約内容を保管。
@@ -331,7 +332,7 @@ export class WWA {
             pathList.push(this._wwaData.mapCGName); //最後に画像ファイル名を追加
             this._wwaData.mapCGName = pathList.join("/"); // pathを復元
             this._restartData = JSON.parse(JSON.stringify(this._wwaData));
-            this._checkOriginalMapString = this._generateMapDataHash(this._restartData);
+            this.checkOriginalMapString = this._generateMapDataHash(this._restartData);
             
             this.initCSSRule();
             this._setProgressBar(getProgress(0, 4, LoadStage.GAME_INIT));
@@ -416,6 +417,9 @@ export class WWA {
             this._scoreWindow = new ScoreWindow(
                 this, new Coord(50, 50), false, util.$id("wwa-wrapper"));
 
+            this._wwaSave = new WWASave(wwa);
+            this._messageWindow.setWWASave(this._wwaSave);
+
             WWACompress.setRestartData(this._restartData);
             this.clearFaces();
             var resumeSaveDataText = util.$id("wwa-wrapper").getAttribute("data-wwa-resume-savedata");
@@ -426,9 +430,13 @@ export class WWA {
                     var resumeData: WWAData = WWACompress.getStartWWAData(resumeSaveDataText);
                     if (this._restartData !== resumeData) {
                         this._applyQuickLoad(resumeData);
-                        this._quickSave(ChoiceCallInfo.CALL_BY_QUICK_SAVE);
+                        this._wwaSave.resumeStart();
                     }
                 }
+            }
+            var autosaveString = util.$id("wwa-wrapper").getAttribute("data-wwa-autosave");
+            if ((autosaveString)&&(autosaveString.match(/^false$/i))) {
+                this._wwaSave.setAutoSaveFlag(false);
             }
             this._setProgressBar(getProgress(2, 4, LoadStage.GAME_INIT));
             this._setLoadingMessage(ctxCover, 4);
@@ -1250,20 +1258,39 @@ export class WWA {
         this._itemMenu.close();
         bg.classList.add("onpress");
         if (button === SidebarButton.QUICK_LOAD) {
-
-            if (this._messageWindow.hasSaveData() && !forcePassword) {
-                this._messageWindow.createSaveDom();
-                if (this._usePassword) {
-
-                    this.setMessageQueue("読み込むデータを選んでください。\n→Ｎｏでデータ復帰用パスワードの\n　入力選択ができます。", true, true);
-                    this._yesNoChoiceCallInfo = ChoiceCallInfo.CALL_BY_QUICK_LOAD;
-                } else {
-
-                    this.setMessageQueue("読み込むデータを選んでください。", true, true);
-                    this._yesNoChoiceCallInfo = ChoiceCallInfo.CALL_BY_QUICK_LOAD;
-                }
-            } else {
-                this.onpasswordloadcalled();
+            this._yesNoChoiceCallInfo = this._wwaSave.getFirstSaveChoiceCallInfo(forcePassword, this._usePassword);
+            switch (this._yesNoChoiceCallInfo) {
+                case ChoiceCallInfo.CALL_BY_QUICK_LOAD:
+                case ChoiceCallInfo.CALL_BY_LOG_QUICK_LOAD:
+                    var secondCallInfo: ChoiceCallInfo;
+                    var loadTest: string = "";
+                    switch (this._yesNoChoiceCallInfo) {
+                        case ChoiceCallInfo.CALL_BY_QUICK_LOAD:
+                            loadTest = "読み込むデータを選んでください。";
+                            this._wwaSave.selectDBSaveDataList();
+                            break;
+                        case ChoiceCallInfo.CALL_BY_LOG_QUICK_LOAD:
+                            loadTest = "読み込むオートセーブを選んでください。";
+                            this._wwaSave.selectLogSaveDataList();
+                            break;
+                    }
+                    secondCallInfo = this._wwaSave.getSecondSaveChoiceCallInfo(this._usePassword);//クイックロード後の選択肢
+                    this._messageWindow.createSaveDom();
+                    switch (secondCallInfo) {
+                        case ChoiceCallInfo.CALL_BY_LOG_QUICK_LOAD:
+                            this.setMessageQueue(loadTest+"\n→Ｎｏでオートセーブ復帰画面に移ります。", true, true);
+                            break;
+                        case ChoiceCallInfo.CALL_BY_PASSWORD_LOAD:
+                            this.setMessageQueue(loadTest +"\n→Ｎｏでデータ復帰用パスワードの\n　入力選択ができます。", true, true);
+                            break;
+                        case ChoiceCallInfo.NONE:
+                            this.setMessageQueue(loadTest, true, true);
+                            break;
+                    }
+                    break;
+                case ChoiceCallInfo.CALL_BY_PASSWORD_LOAD:
+                    this.onpasswordloadcalled();
+                    break;
             }
         } else if (button === SidebarButton.QUICK_SAVE) {
             if (!this._wwaData.disableSaveFlag) {
@@ -1405,7 +1432,12 @@ export class WWA {
 
 
         if (this._player.isControllable()) {
-
+            if (!this._wwaData.disableSaveFlag) {
+                //オートセーブ処理
+                if (this._wwaSave.isAutoSaveFrame(this._player)) {
+                    this._quickSave(ChoiceCallInfo.CALL_BY_LOG_QUICK_SAVE)
+                }
+            }
             if (this._player.isDelayFrame()) {
                 /*
                  * メッセージ表示直後、何も操作しないフレームを用意する。
@@ -2790,7 +2822,8 @@ export class WWA {
                     }
                 } else if (this._yesNoChoiceCallInfo === ChoiceCallInfo.CALL_BY_ITEM_USE) {
                     this._player.readyToUseItem(this._yesNoUseItemPos);
-                } else if (this._yesNoChoiceCallInfo === ChoiceCallInfo.CALL_BY_QUICK_LOAD) {
+                } else if ((this._yesNoChoiceCallInfo === ChoiceCallInfo.CALL_BY_QUICK_LOAD) ||
+                           (this._yesNoChoiceCallInfo === ChoiceCallInfo.CALL_BY_LOG_QUICK_LOAD)) {
                     this._messageWindow.deleteSaveDom();
                     (<HTMLDivElement>(util.$id(sidebarButtonCellElementID[SidebarButton.QUICK_LOAD]))).classList.remove("onpress");
 
@@ -2859,14 +2892,32 @@ export class WWA {
                 } else if (this._yesNoChoiceCallInfo === ChoiceCallInfo.CALL_BY_ITEM_USE) {
                     var bg = <HTMLDivElement>(util.$id("item" + (this._yesNoUseItemPos - 1)));
                     bg.classList.remove("onpress");
-                } else if (this._yesNoChoiceCallInfo === ChoiceCallInfo.CALL_BY_QUICK_LOAD) {
+                } else if ((this._yesNoChoiceCallInfo === ChoiceCallInfo.CALL_BY_QUICK_LOAD) ||
+                           (this._yesNoChoiceCallInfo === ChoiceCallInfo.CALL_BY_LOG_QUICK_LOAD)) {
                     this._messageWindow.deleteSaveDom();
-                    if (this._usePassword) {
-                        this._yesNoJudge = YesNoState.UNSELECTED;
-                        this.onpasswordloadcalled();
-                        return;
-                    } else {
-                        (<HTMLDivElement>(util.$id(sidebarButtonCellElementID[SidebarButton.QUICK_LOAD]))).classList.remove("onpress");
+                    this._yesNoChoiceCallInfo = this._wwaSave.getSecondSaveChoiceCallInfo(this._usePassword);
+                    switch (this._yesNoChoiceCallInfo) {
+                        case ChoiceCallInfo.NONE:
+                            (<HTMLDivElement>(util.$id(sidebarButtonCellElementID[SidebarButton.QUICK_LOAD]))).classList.remove("onpress");
+                        case ChoiceCallInfo.CALL_BY_PASSWORD_LOAD:
+                            this._yesNoJudge = YesNoState.UNSELECTED;
+                            this.onpasswordloadcalled();
+                            return;
+                        case ChoiceCallInfo.CALL_BY_LOG_QUICK_LOAD:
+                            this._wwaSave.selectLogSaveDataList();
+                            this._messageWindow.createSaveDom();
+                            var secondCallInfo: ChoiceCallInfo = this._wwaSave.getSecondSaveChoiceCallInfo(this._usePassword);//クイックロード後の選択肢
+                            this._yesNoJudge = YesNoState.UNSELECTED;
+                            switch (secondCallInfo) {
+                                case ChoiceCallInfo.CALL_BY_PASSWORD_LOAD:
+                                    this.setMessageQueue("読み込むオートセーブを選んでください。\n→Ｎｏでデータ復帰用パスワードの\n　入力選択ができます。", true, true);
+                                    break;
+                                case ChoiceCallInfo.NONE:
+                                    this.setMessageQueue("読み込むオートセーブを選んでください。", true, true);
+                                    break;
+                            }
+
+                            return;
                     }
                 } else if (this._yesNoChoiceCallInfo === ChoiceCallInfo.CALL_BY_QUICK_SAVE) {
 
@@ -3432,10 +3483,11 @@ export class WWA {
 
         switch (callInfo) {
             case ChoiceCallInfo.CALL_BY_QUICK_SAVE:
+            case ChoiceCallInfo.CALL_BY_LOG_QUICK_SAVE:
                 qd.checkString = this._generateSaveDataHash(qd);
                 break;
             case ChoiceCallInfo.CALL_BY_PASSWORD_SAVE:
-                qd.checkOriginalMapString = this._checkOriginalMapString;
+                qd.checkOriginalMapString = this.checkOriginalMapString;
                 qd.mapCompressed = this._compressMap(qd.map);
                 qd.mapObjectCompressed = this._compressMap(qd.mapObject);
                 qd.checkString = this._generateSaveDataHash(qd);
@@ -3479,6 +3531,13 @@ export class WWA {
                     compress: WWACompress.compress(qd)
                 });
                 break;
+            case ChoiceCallInfo.CALL_BY_LOG_QUICK_SAVE:
+                this.wwaCustomEvent('wwa_autosave', {
+                    data: qd,
+                    compress: WWACompress.compress(qd)
+                });
+                this._wwaSave.autoSave(this._cvs, qd);
+                break;
         }
     }
 
@@ -3497,7 +3556,7 @@ export class WWA {
     }
 
     private _decodePassword(pass: string): WWAData {
-        var ori = this._checkOriginalMapString;
+        var ori = this.checkOriginalMapString;
         try {
             var json = CryptoJS.AES.decrypt(
                 pass,
@@ -3516,7 +3575,7 @@ export class WWA {
     }
 
     private _quickLoad(restart: boolean = false, password: string = null, apply: boolean = true): WWAData {
-        if (!restart && this._messageWindow.hasSaveData() === void 0 && password === null) {
+        if (!restart && this._wwaSave.hasSaveData() === void 0 && password === null) {
             throw new Error("セーブデータがありません。");
         }
         var newData: WWAData;
@@ -3544,7 +3603,7 @@ export class WWA {
                 throw new Error("データが壊れているようです。\nInvalid hash (ALL DATA)= " + newData.checkString + " " + this._generateSaveDataHash(newData));
             }
 
-            var checkOriginalMapString = this._checkOriginalMapString;
+            var checkOriginalMapString = this.checkOriginalMapString;
             if (newData.checkOriginalMapString !== checkOriginalMapString) {
                 throw new Error("管理者によってマップが変更されたようです。\nInvalid hash (ORIGINAL MAP)= " + newData.checkString + " " + this._generateSaveDataHash(newData));
             }
@@ -3584,6 +3643,7 @@ export class WWA {
         this._mapIDTableCreate();
         this._replaceAllRandomObjects();
         this.updateCSSRule();
+        this._wwaSave.setPlayer(this._player);
     }
     private _mapIDTableCreate(): void {
         var pid: number;
