+++ packages/engine/src/wwa_save/WWASaveDataLog.ts
@@ -1,0 +1,4 @@
+import WWASaveData from "./WWASaveData";
+
+export default class WWASaveDataLog extends WWASaveData {
+}
